Added play for docker registry; improved others
A play for deploying a private docker registry in Kubernetes has been
added.

The play deploys a rook cluster first because for persistence a
persistent volume system is needed. At the end it also loads and pushs
the images that are available in a sub-folder "docker-images".

The plays for loading and pushing images and the rook cluster have been
improved with tags. The play for rook cluster also includes now a pause
to ensure it's ready for use after the play (yet a better way has to be
found).